Hoist hts-forking package to repo root
