--- @hts-forking/resources/HtsSystemContract.json
@@ -1,1 +1,0 @@
-../../out/HtsSystemContract.sol/HtsSystemContract.json
