+++ package.json
@@ -1,30 +1,50 @@
 {
-    "name": "@hashgraph/hardhat-forking-plugin",
+    "name": "@hashgraph/system-contracts-forking",
     "version": "0.0.1",
-    "description": "Hedera HTS Forking Support",
-    "exports": "./@hardhat-forking-plugin/src/index.js",
-    "private": true,
-    "workspaces": [
-        "@hts-forking",
-        "@hardhat-forking-plugin"
+    "description": "Hashgraph System Contracts and Hardhat plugin to support fork testing",
+    "repository": "github:hashgraph/hedera-forking",
+    "author": "2024 Hedera Hashgraph, LLC",
+    "license": "Apache-2.0",
+    "keywords": [
+        "ethereum",
+        "evm",
+        "hardhat",
+        "hardhat-plugin",
+        "hashgraph",
+        "hedera",
+        "hedera-token-service",
+        "precompiles",
+        "smart-contracts",
+        "system-contracts"
     ],
+    "files": [
+        "src/",
+        "out/HtsSystemContract.sol/"
+    ],
+    "exports": "./src/index.js",
     "scripts": {
         "prepare": "husky",
         "type-check": "tsc",
         "lint": "eslint",
+        "test": "mocha",
         "make:readme": "scripts/readme.js README.md"
     },
-    "author": "2024 Hedera Hashgraph, LLC",
-    "license": "Apache-2.0",
+    "dependencies": {
+        "ethers": "^6.1.0"
+    },
     "devDependencies": {
         "@eslint/js": "^9.13.0",
         "@hashgraph/sdk": "^2.54.2",
+        "@types/chai": "^4.3.20",
+        "@types/mocha": "^10.0.7",
+        "chai": "^4.0.0",
         "eslint": "^9.13.0",
         "eslint-plugin-license-header": "^0.6.1",
         "eslint-plugin-mocha": "^10.5.0",
         "globals": "^15.11.0",
         "husky": "^9.1.5",
         "lint-staged": "^15.2.10",
+        "mocha": "^10.7.3",
         "prettier": "^3.3.3",
         "typescript": "^5.4.5",
         "typescript-eslint": "^8.12.0"
