+++ test/data/MFCT/getBalanceOfToken_0.0.2669675.json
