+++ test/data/MFCT/getTokenRelationship_0xa3612a87022a4706fc9452c50abd2703ac4fd7d9.json
