+++ test/data/USDC/getBalanceOfToken_0.0.1421.json
