+++ test/data/USDC/getBalanceOfToken_0.0.2183.json
