+++ test/data/USDC/getBalanceOfToken_0.0.5176.json
