+++ test/data/USDC/getToken.json
