+++ test/data/USDC/getTokenRelationship_0.0.5176.json
